restore original notes section and move ui to another file #15
